6.40. Сборка сцены «Замочная скважина»
